refactor(path): update path
